+++ R/dplyr-rename.R
@@ -2,7 +2,7 @@
 #'
 #' @description
 #' These two functions provide a way to rename columns in the sample or variable
-#' information tibble of an [experiment()].
+#' information of an [experiment()] or `SummarizedExperiment`.
 #'
 #' The same syntax as `dplyr::rename()` is used.
 #' For example, to rename the "group" column in the sample information tibble to "condition",
@@ -12,12 +12,13 @@
 #' These two columns are used to link the sample or variable information tibble
 #' to the expression matrix.
 #'
-#' @param exp An [experiment()].
+#' @param exp An [experiment()] or `SummarizedExperiment` object.
 #' @param ... <[`data-masking`][rlang::args_data_masking]> Name pairs to rename.
 #'   Use `new_name = old_name` to rename columns.
 #'
-#' @returns An new [experiment()] object.
+#' @returns An object of the same class as `exp`.
 #'
+#' @inheritSection mutate_obs Identifier columns
 #' @examples
 #' toy_exp <- toy_experiment
 #' toy_exp
@@ -51,11 +52,16 @@ rename_var <- function(exp, ...) {
 
 # Internal function that handles the common logic for both rename_obs and rename_var
 rename_info_data <- function(exp, info_field, id_column, ...) {
-  stopifnot(is_experiment(exp))
+  stopifnot(is_tidy_container(exp))
+  id_column <- tidy_id_column(exp, id_column)
 
   # Get original data and rename it
-  original_data <- exp[[info_field]]
+  original_data <- tidy_info_data(exp, info_field, id_column)
   new_data <- rename_data(original_data, info_field, id_column, ...)
+
+  if (methods::is(exp, "SummarizedExperiment")) {
+    return(update_se_info(exp, new_data, info_field, id_column))
+  }
 
   # Create new experiment object
   new_exp <- exp
@@ -65,20 +71,29 @@ rename_info_data <- function(exp, info_field, id_column, ...) {
 }
 
 rename_data <- function(data, data_name, info_type, ...) {
+  protected_columns <- intersect(
+    c(info_type, tidy_position_column()),
+    colnames(data)
+  )
+
   # Create a prototype (empty data frame with same structure) for validation
   prototype <- data[0, ]
 
-  # Remove the ID column from prototype for validation
-  prototype_without_id <- dplyr::select(prototype, -dplyr::all_of(info_type))
+  # Remove protected columns from prototype for validation
+  prototype_without_id <- dplyr::select(
+    prototype,
+    -dplyr::all_of(protected_columns)
+  )
 
   # Try renaming on the prototype first to validate
   validate_rename(prototype_without_id, data_name, info_type, ...)
 
   # If validation passes, perform the actual renaming
-  index_col <- data[[info_type]]
-  new_data <- dplyr::select(data, -dplyr::all_of(info_type))
+  protected_data <- dplyr::select(data, dplyr::all_of(protected_columns))
+  new_data <- dplyr::select(data, -dplyr::all_of(protected_columns))
   new_data <- dplyr::rename(new_data, ...)
-  new_data <- dplyr::mutate(new_data, "{info_type}" := index_col, .before = 1)
+  check_reserved_tidy_columns(new_data, protected_columns, data_name)
+  new_data <- dplyr::bind_cols(protected_data, new_data)
 
   new_data
 }