+++ tests/testthat/test-dplyr-rename.R
@@ -8,6 +8,41 @@ test_that("renaming info tibbles works", {
   expect_identical(colnames(exp2$var_info), c("variable", "new_type"))
 })
 
+test_that("rename verbs support SummarizedExperiment", {
+  se <- create_test_se(c("S1", "S2", "S3"), c("V1", "V2", "V3"))
+  SummarizedExperiment::colData(se)$sample <- c("A", "B", "C")
+  SummarizedExperiment::rowData(se)$variable <- c("X", "Y", "Z")
+
+  result <- se |>
+    rename_obs(condition = group, sample_id = sample) |>
+    rename_var(class = type, variable_id = variable)
+
+  expect_identical(
+    colnames(SummarizedExperiment::colData(result)),
+    c("condition", "sample_id")
+  )
+  expect_identical(
+    colnames(SummarizedExperiment::rowData(result)),
+    c("class", "variable_id")
+  )
+  expect_identical(colnames(result), c("S1", "S2", "S3"))
+  expect_identical(rownames(result), c("V1", "V2", "V3"))
+
+  expect_error(rename_obs(se, id = .sample), "could not rename the.*\\.sample")
+  expect_error(
+    rename_var(se, id = .variable),
+    "could not rename the.*\\.variable"
+  )
+  expect_error(
+    rename_obs(se, .sample = sample),
+    "reserved for dimension names"
+  )
+  expect_error(
+    rename_var(se, .variable = variable),
+    "reserved for dimension names"
+  )
+})
+
 
 test_that("trying to rename non-existing columns throws an error", {
   exp <- create_test_exp(c("S1", "S2", "S3"), c("V1", "V2", "V3"))